backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Environment variables
python-dotenv==1.0.0

# OpenAGI Lux (adjust version as needed)
oagi>=0.1.0

# HTTP client
httpx==0.25.2

# Logging and monitoring
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1